requirements.txt
python-telegram-bot==20.4
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
pytz>=2024.1

